+++ CHANGES.md
@@ -22,6 +22,7 @@
 - Allow three character hex and web colors in color picker hex input [#1184](https://github.com/penpot/penpot/issues/1184)
 - Allow lowercase search for fonts [#1180](https://github.com/penpot/penpot/issues/1180)
 - Fix group renaming problem [Taiga #1969](https://tree.taiga.io/project/penpot/issue/1969)
+- Fix export group with shadows on children [Taiga #2036](https://tree.taiga.io/project/penpot/issue/2036)
 
 ### :arrow_up: Deps updates
 ### :boom: Breaking changes